requirements.txt
anyio==3.7.0
async-generator==1.10
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
exceptiongroup==1.1.1
fake-useragent==1.1.3
fastapi==0.99.1
ghp-import==2.1.0
greenlet==2.0.2
griffe==0.27.0
h11==0.14.0
idna==3.4
Jinja2==3.1.2
Markdown==3.3.7
MarkupSafe==2.1.2
mergedeep==1.3.4
mkdocs==1.4.2
mkdocs-autorefs==0.4.1
mkdocs-glightbox==0.3.2
mkdocs-material==9.1.6
mkdocs-material-extensions==1.1.1
mkdocstrings==0.21.2
mkdocstrings-python==0.9.0
mysql-connector-python==8.0.33
numpy==1.25.0
outcome==1.2.0
packaging==23.1
protobuf==3.20.3
pydantic==1.10.9
Pygments==2.15.0
pymdown-extensions==9.11
PyMySQL==1.1.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
pyyaml_env_tag==0.1
regex==2023.3.23
requests==2.31.0
selenium==4.10.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
SQLAlchemy==2.0.17
starlette==0.27.0
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.3
typing_extensions==4.7.0
tzdata==2023.3
urllib3==2.0.3
uvicorn==0.22.0
watchdog==3.0.0
webdriver-manager==3.8.6
wsproto==1.2.0
